Adding missing offsets; it's not really necessary to look at TIOFeatures, though one could.

+++ uproot4/const.py
@@ -102,4 +102,4 @@ kSTLany = 300
 
 ############# IOFeatures
 
-kGenerateOffsetMap = 1
+kGenerateOffsetMap = numpy.uint8(1)
